refactor: remove superseded persistence machinery
Closes #379
