+++ docs/agents/architecture.md
@@ -9,8 +9,8 @@ These rules apply to runtime and type-only imports:
 - `entrypoints/` owns WXT registration, background message execution, popup UI, and query hooks.
 - `content/` owns LeetCode DOM/GraphQL integration. `content/ui/` owns presentation; `content/rating-actions.ts` owns problem lookup and background RPC orchestration; `content/bootstrap.tsx` owns mounting and lifecycle.
 - `domain/` owns application models and schemas, scheduling, review days, settings, and language. Pass explicit inputs; keep browser, storage, service, messaging, and translation-catalog dependencies out. `ts-fsrs` is allowed.
-- `services/` owns workflows, clocks, settings reads, FSRS lifetime, write order, and sync decisions. Services may compose domain rules, adapters, and other services without cycles. Statistics and editor reset read cards through persistence rather than the card service.
-- `infrastructure/` owns storage keys, codecs, migrations, backup-format parsing and relationship checks, GitHub requests, browser messaging, and language detection. It must not depend on services or UI.
+- `services/` owns workflows, clocks, settings reads, FSRS lifetime, write order, and sync decisions. Services may compose domain rules, adapters, and other services without cycles. Learning workflows, including statistics and editor reset, use one captured document through the storage entry point.
+- `infrastructure/` owns storage keys, codecs, document conversions, backup-format parsing and relationship checks, GitHub requests, browser messaging, and language detection. It must not depend on services or UI.
 - `i18n/` owns translation catalogs and the `Translations` type. It may import the domain language type, but has no browser, storage, service, or UI dependencies. Settings policy uses the domain language registry without loading dictionaries.
 - Popup and content workflows use `infrastructure/browser/messages.ts` rather than importing services or persistence. Content may use the read-only `infrastructure/storage/translations.ts` adapter.
 - The popup owns permissions, active-tab inspection, and banner dismissal state. Permission requests originate from user interactions.
@@ -30,13 +30,13 @@ Before changing background commands or sync execution, read [ADR-0001](../adr/00
 
 ## Storage, backup, and sync
 
-[ADR-0003](../adr/0003-single-document-learning-data.md) records the single-document design. The document is authoritative at runtime; the [canonical implementation plan](https://github.com/mattcdrake/LeetSRS/issues/371) tracks the remaining superseded-code cleanup.
+[ADR-0003](../adr/0003-single-document-learning-data.md) records the single-document design. The document is authoritative at runtime. The superseded migration lifecycle and collection/settings persistence wrappers have been removed; historical compatibility lives in the startup bridge and pure document conversions.
 
 - One browser-local document owns its schema version, modification timestamp, slug-keyed cards with embedded notes, daily statistics, and stored settings overrides. Dates remain numbers. Missing settings retain normal defaults, including browser-language detection.
 - Domain owns current schemas and explicit-input policy. Services prepare and validate complete changes, including timestamps, before infrastructure writes the document once. Reviews commit cards, statistics, and their edit timestamp together; queue calculations use one captured document and time. Keep normal operations behind startup readiness and the existing background write queue.
-- Migrations are pure document-version conversions shared by startup and backup import. Preserve supported historical versions and their meaning, frozen historical data contracts, and precedence rules. Validate historical input before stripping extras and validate the current result. Keep browser storage, clocks, per-step loaders/savers/cleanup, physical-layout modules, and intermediate completion writes out of the conversion sequence.
+- `infrastructure/storage/learning-document.ts` owns reads, replacement, backup parsing, and pure document-version conversion orchestration shared by startup and import. `document-conversions/` owns the frozen historical contracts and individual conversions. Preserve supported historical versions and their meaning, frozen historical data contracts, and precedence rules. Validate historical input before stripping extras and validate the current result. Keep browser storage, clocks, per-step loaders/savers/cleanup, physical-layout modules, and intermediate completion writes out of the conversion sequence.
 - One startup bridge gathers scattered legacy values only when the new document is absent. Prepare and validate before writing the complete document and its version. After success the document is authoritative; cleanup failure must not trigger re-import, and corrupt or future documents must not fall back to stale legacy keys. Preserve credentials and already-combined connection values over retained legacy fields.
 - File backups and Gist payloads use the document representation. Import validates and converts before one replacement, preserves the imported timestamp, and clears omitted notes/settings through replacement. Retain legacy backup timestamp fallback without treating conversion as a learner edit.
 - The PAT, Gist ID, and automatic-sync setting remain a separate browser-synced connection. Sync status stays local. Import leaves both untouched; explicit Reset All Data also clears the connection, retained legacy connection keys, and local status.
 - Whole-dataset conflict resolution remains [ADR-0002](../adr/0002-whole-dataset-gist-sync.md). Reject unsupported remote formats before replacing data or writing a Gist. All browsers sharing a Gist must update before resuming sync; the transition does not support dual writes or old and new clients syncing together.
-- Preserve supported learning data and meaningful failure behavior. The implementation must remove the superseded migration/persistence machinery and sequencing-only tests. A single document does not imply a transaction across connection/status writes or a crash-recovery framework.
+- Preserve supported learning data and meaningful failure behavior. Keep obsolete migration/persistence machinery and sequencing-only tests out of the runtime and test suite. A single document does not imply a transaction across connection/status writes or a crash-recovery framework.
