--- infrastructure/storage/data-tracker.ts
@@ -1,5 +1,0 @@
-import { writeSyncMetadata } from './sync-metadata';
-
-export async function markDataUpdated(): Promise<void> {
-  await writeSyncMetadata('dataUpdatedAt', new Date().toISOString());
-}
