+++ services/settings.ts
@@ -1,48 +1,32 @@
-import {
-  DEFAULT_SETTINGS,
-  SETTING_KEYS,
-  type Settings,
-  type SettingsUpdate,
-  settingsSchema,
-  settingsUpdateSchema,
-} from '@/domain/settings';
+import { learningDocumentSchema } from '@/domain/learning-document';
+import { resolveSettings, type Settings, type SettingsUpdate, settingsUpdateSchema } from '@/domain/settings';
 import { detectBrowserLanguage } from '@/infrastructure/browser/language';
-import { markDataUpdated } from '@/infrastructure/storage/data-tracker';
-import { readSetting, removeSetting, writeSetting } from '@/infrastructure/storage/settings';
+import { readLearningDocument, replaceLearningDocument } from '@/infrastructure/storage/learning-document';
 
 export async function getSettings(): Promise<Settings> {
-  const entries = await Promise.all(
-    SETTING_KEYS.map(async (key) => {
-      const value = await readSetting(key);
-      return [key, value ?? (key === 'language' ? detectBrowserLanguage() : DEFAULT_SETTINGS[key])] as const;
-    })
-  );
-  return settingsSchema.parse(Object.fromEntries(entries));
+  const document = await readLearningDocument();
+  if (!document) {
+    throw new Error('Learning document is not initialized');
+  }
+
+  return resolveSettings(document.settings, document.settings.language ?? detectBrowserLanguage());
 }
 
 export async function updateSettings(changes: SettingsUpdate): Promise<void> {
   const parsedChanges = settingsUpdateSchema.parse(changes);
-  const changedKeys = SETTING_KEYS.filter((key) => Object.hasOwn(parsedChanges, key));
-
-  if (changedKeys.length === 0) {
+  if (Object.keys(parsedChanges).length === 0) {
     return;
   }
 
-  await Promise.all(changedKeys.map((key) => writeSetting(key, parsedChanges[key] as Settings[typeof key])));
-  await markDataUpdated();
-}
+  const document = await readLearningDocument();
+  if (!document) {
+    throw new Error('Learning document is not initialized');
+  }
 
-export async function exportSettings(): Promise<Partial<Settings>> {
-  const entries = await Promise.all(
-    SETTING_KEYS.map(async (key) => {
-      const value = await readSetting(key);
-      return value !== null ? ([key, value] as const) : null;
-    })
-  );
-
-  return Object.fromEntries(entries.filter((entry) => entry !== null)) as Partial<Settings>;
-}
-
-export async function resetSettings(): Promise<void> {
-  await Promise.all(SETTING_KEYS.map((key) => removeSetting(key)));
+  const next = learningDocumentSchema.parse({
+    ...document,
+    settings: { ...document.settings, ...parsedChanges },
+    dataUpdatedAt: new Date().toISOString(),
+  });
+  await replaceLearningDocument(next);
 }
